pyproject.toml
[build-system]
requires = ["setuptools >= 61.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rl-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "jax>=0.4.20",
    "mlflow>=2.8.0",
    "numpy>=1.24.0",
    "gin-config>=0.5.0",
    "classic-pacmdp-envs",
    "minimal-agents",
]

[project.scripts]
rl-run = "rl_research.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "ruff>=0.2.0",
]

[tool.black]

[tool.isort]
atomic = true
profile = "black"
src_paths = ["rl_research/"]
extra_standard_library = ["typing_extensions"]
indent = 4
lines_after_imports = 2
multi_line_output = 3

[tool.setuptools]
py-modules = ["rl_research"]

[tool.pyright]
include = ["rl_research/**"]
exclude = ["tests/**", "**/node_modules", "**/__pycache__", "outputs/**"]
strict = []

typeCheckingMode = "basic"
pythonVersion = "3.11"
pythonPlatform = "All"
typeshedPath = "typeshed"
enableTypeIgnoreComments = true

[tool.uv.sources]
classic-pacmdp-envs = { git = "https://github.com/cruz-lucas/classic_pacmdp_envs.git" }
minimal-agents = { git = "https://github.com/cruz-lucas/minimal_agents.git" }
